moved implementation cpp added explicit inst
fixed bug for float
